+++ ct-app/core/components/__init__.py
@@ -1,10 +1,9 @@
-from . import decorators
+from . import config_parser, decorators
 from .address import Address
 from .asyncloop import AsyncLoop
 from .environment_utils import EnvironmentUtils
 from .lockedvar import LockedVar
 from .messages import MessageFormat, MessageQueue
-from .parameters import Parameters
 from .peer import Peer
 from .singleton import Singleton
 from .utils import Utils
@@ -18,7 +17,7 @@ __all__ = [
     "LockedVar",
     "MessageQueue",
     "MessageFormat",
-    "Parameters",
+    "config_parser",
     "Singleton",
     "Utils",
     "decorators",
