+++ ct-app/core/components/config_parser/funding.py
@@ -1,0 +1,5 @@
+from .base_classes import ExplicitParams
+
+
+class FundingParams(ExplicitParams):
+    keys = {"constant": float}
