+++ ct-app/core/components/config_parser/peer.py
@@ -1,0 +1,10 @@
+from .base_classes import ExplicitParams
+
+
+class PeerParams(ExplicitParams):
+    keys = {
+        "min_version": str,
+        "sleep_mean_time": int,
+        "sleep_std_time": int,
+        "message_multiplier": int,
+    }
