+++ ct-app/core/node.py
@@ -2,14 +2,14 @@
 import logging
 from datetime import datetime
 
-from prometheus_client import Gauge, Histogram
-
 from core.components.asyncloop import AsyncLoop
 from core.components.logs import configure_logging
+from prometheus_client import Gauge, Histogram
 
 from .api import HoprdAPI
-from .components import LockedVar, Parameters, Peer, Utils
+from .components import LockedVar, Peer, Utils
 from .components.address import Address
+from .components.config_parser import Parameters
 from .components.decorators import connectguard, keepalive, master
 from .components.messages import MessageFormat, MessageQueue
 from .components.node_helper import NodeHelper
