+++ ct-app/test/components/test_code_for_environment/test_main2.py
@@ -1,4 +1,4 @@
-from core.components import Parameters
+from core.components.config_parser import Parameters
 
 
 def main():
