+++ ct-app/test/economic_model/test_economic_model_legacy.py
@@ -1,7 +1,6 @@
 import pytest
-
 from core.api.response_objects import TicketPrice
-from core.components.parameters import LegacyParams
+from core.components.config_parser.economic_model import LegacyParams
 
 
 @pytest.fixture
